construction.jl

Ce module est le programme auxiliaire de l'application.

Il prend en entrée la série temporelle x au format binaire Int16 et génère xQ, sa version sous-quantifiée à 1 bit, et P, l'histogramme des couples de valeurs successives de x.

xQ est sauvegardé dans un fichier .dat et P est sauvegardé dans un fichier .ppm.

Ces deux fichiers sont ensuite utilisés par le programme principal comme entrée.

Ce module n'a aucune dépendance interne ni externe.

API

L2F2_Dequantification_App.construire_p — Method

construire_p(Vector{Int16})::Dict{Tuple{Int16, Int16}, Int}

Construit P sous la forme d'un dictionnaire, avec comme clé le couple x[n-1], x[n], et comme valeur son nombre d’apparition.

Paramètres:

série temporelle x.

L2F2_Dequantification_App.lire_p — Method

lire_p(String):: Dict{Tuple{Int16,Int16},Int}

Construit la représentation du fichier qui contient P sous la forme d'un dictionnaire.

Paramètres:

fichier qui contient P

L2F2_Dequantification_App.lire_serie — Method

lire_serie(String)::Vector{Int16}

Prend en entrée le chemin de x et stock les données dans un vecteur.

Vérifie si le fichier est vide.

Paramètres:

chemin

L2F2_Dequantification_App.sauvegarder_p — Method

sauvegarder_p(Dict{Tuple{Int16, Int16}, Int} x String)::Nothing

Sauvegarde le dictionnaire P dans un fichier.

Paramètres:

chemin du fichier de sauvegarde.

L2F2_Dequantification_App.sauvegarder_xq — Method

sauvegarder_xq(Vector{Int16} x String)::Nothing

Sauvegarde la série sous-quantifiée xQ dans un fichier.

Paramètres:

chemin du fichier de sauvegarde.

L2F2_Dequantification_App.sous_quantifier — Method

sous_quantifier(Vector{Int16})::Vector{Int16}

Transforme chaque valeur de la série temporelle x construit avec lire_serie() en l'entier pair inférieur ou égal.

Autrement dit, construit xQ.

Paramètres:

série temporelle x