interface.jl

Ce module gère l'interface graphique de l'application.

Dépendances externes

GLMakie: gère l'interface graphique: boutons, layout...

GraphMakie: gère la visualisation de l'arbre binaire.

NativeFileDialog: ouvre le sélecteur de fichiers natif du système.

ZipFile: permet l'export des solutions dans une archive ZIP

L'interface se divise en deux colonnes. La colonne gauche affiche la liste des solutions trouvées, le bouton d'export et le bouton télécharger xQ et P. La colonne droite affiche l'arbre en cours de construction ainsi que les boutons de navigation : Importer x, Lancer, Arrêter et Importer xQ et P.

Le nombre de branches restantes est affiché en temps réel et mis à jour à chaque appel de la fonction developper du module déquantification.jl.

Ce module dépend de dequantification.jl et arbre.jl.

API

L2F2_Dequantification_App.a_descendant_vivant — Method

a_descendant_vivant(Noeud)::Bool

Retourne true si le noeud est vivant ou a au moins un descendant vivant. Sert à colorier en bleu tous les noeuds sur le chemin d'une branche vivante.

Paramètres :

n : noeud à tester

L2F2_Dequantification_App.affiche_erreur — Method

affiche_erreur(Axis x String x Ref):: Nothing

Affiche un message d'erreur à l'ecran, dans l'emplacement (=Axis) du graphe.

Paramètres:

l'axe du graphe dans lequel va s'afficher le message d'erreur.

le message.

la référence du texte d'erreur

L2F2_Dequantification_App.affiche_succes — Method

affiche_succes(Observable{String} x String)::Nothing

Affiche un message de succès dans le cas où l'importation ou l'exportation des fichiers/dossiers réussit

Paramètres:

observable du message de réussite

message de réussite

L2F2_Dequantification_App.arreter_animation — Method

arreter_animation(Observable{Bool})::Nothing

Arrête l'animation si l'utilisateur clique sur le bouton arrêter.

Paramètres:

la référence est_lance qui permet de prévenir l'algorithme que l'utilisateur a cliqué sur le bouton d'arrêt.

L2F2_Dequantification_App.calculer_profondeur! — Method

calculer_profondeur!(Noeud x Dict{Noeud,Int} x Int)::Nothing

Calcule la profondeur de chaque noeud dans l'arbre pour déterminer sa position verticale. La racine est à la profondeur 0, ses enfants sont à la profondeur 1 etc.

Paramètres :

noeud : noeud courant à traiter

profondeurs : dictionnaire associant chaque noeud à sa profondeur

prof : profondeur courante

L2F2_Dequantification_App.calculer_x! — Method

calculer_x!(Noeud x Dict{Noeud,Float32} x Ref{Int})::Nothing

Calcule la position horizontale de chaque noeud pour l'affichage. Les feuilles sont placées de gauche à droite. Les autres noeuds sont au centre entre leurs enfants.

Paramètres :

noeud : noeud courant à traiter

x_pos : dictionnaire associant chaque noeud à sa position horizontale

compteur : compteur partagé incrémenté à chaque feuille rencontrée

L2F2_Dequantification_App.convertir_arbre — Method

convertir_arbre(Arbre)::(SimpleDiGraph, Vector{Point2f})

Convertit la structure Arbre en un SimpleDiGraph.

Paramètres :

arbre : l'arbre à convertir

Retourne :

le graphe orienté représentant l'arbre

un vecteur avec les positions de chaque noeud.

un vecteur avec la couleur de chaque sommet (bleu si vivant ou ancetre vivant, rouge sinon)

L2F2_Dequantification_App.creer_interface — Method

creer_interface()::Nothing

Initialise la fenêtre et réinitialise le dossier de sauvegarde des solutions trouvées lors de la dérnière session. Initialise tous les observables définis. Crée les layout. Définit le comportement lié aux boutons.

L2F2_Dequantification_App.formater_temps — Method

formater_temps(ms::Float64)::String

Convertit une durée en millisecondes dans le format (ms, s, min, h).

Paramètres:

ms : durée en millisecondes qu'on va formater.

L2F2_Dequantification_App.importer_fichier — Method

importer_fichier(Axis x Ref x Observable{String}):: Vector{Int16}

Ouvre l'explorateur de fichiers natif et charge le fichier x séléctionné. En cas de réussite, affiche un message de réussite, sinon affiche un message d'erreurs dans l'emplacement du graphe.

Paramètres:

on passe en entrée l'axe du graphe afin de gérer les erreur

référence texte_erreur

observable message_succès

L2F2_Dequantification_App.importer_xQP — Method

importer_xQP(Axis x Ref{Union{Nothing, Vector{Int16}}} x Ref{Union{Nothing, Dict{Tuple{Int16,Int16}, Int}}} x Ref{Any} x Observable{String}, Observable{String})::Nothing

Importe le dossier contenant xQ et P et indique à l'interface si le chargement a réussi. Affiche un message d'erreur en cas d'erreur (fichiers non trouvables) ou un message de réussite en cas de succès.

Paramètres:

l'axe dans lequel seront affiché les messages d'erreurs.

les références xQ_charge et P_charge permettant de notifier l'interface (notamment lancer_animation) que les données ont bien été chargé.

référence du texte d'erreur

observable du message de réussite

observable du message à l'écran confirmant le nom du fichier actif

L2F2_Dequantification_App.indexer_noeuds! — Method

indexer_noeuds!(SimpleDiGraph x Noeud x Dict{Noeud,Int})::Nothing

Parcourt l'arbre en profondeur et donne identifiant unique à chaque noeud.

Paramètres :

graphe : le graphe auquel on ajoute les noeuds

noeud : noeud courant à indexer

indices : dictionnaire associant chaque noeud à son identifiant dans le graphe

L2F2_Dequantification_App.lancer_animation — Method

lancer_animation(Axis x Ref{Any}xRef{Union{Nothing, Vector{Int16}}} x Ref{Union{Nothing, Dict{Tuple{Int16,Int16}, Int}}} x Observable{Bool} x Observable{SimpleDiGraph} x Observable{Float64}xObservable{Int} x Observable{Vector{String}}x FunctionxFunctionxString x Ref{Bool})::Nothing

Vérifie que les données sont chargées, réinitialise l'interface graphique, met à jour le dossier de sauvegarde et lance l'animation.

Paramètres:

l'axe dans lequel vit le graphe,

la référence du texte d'erreur qui s'affichera à la place de l'arbre,

les références vers xQ et P,

l'observable est_lance indiquant que l'animation est en cours,

l'observable graphe_obs contenant l'arbre et se mettant à jour au fur et à mesure de la déquantification,

l'observable branche_obs mettant à jour le nombre de branches restantes,

l'observable solutions_obs stockant toutes les solutions trouvées,

les fonctions misajourarbre, misajourbranches,

le dossier de sauvegarde des solutions,

la référence est_lance qui permet de mettre la reference continuer à false si l'utilisateur clique sur arrêter.

Note: continuer est la référence permettant de prévenir l'algorithme que l'utilisateur a cliqué sur arrêter.

L2F2_Dequantification_App.mis_a_jour_arbre — Method

mis_a_jour_arbre(Observable x Observable x Observable x Arbre)::Nothing

Met à jour les observables du graphe et ses postions.

Paramètres :

graphe_obs : observable du graphe affiché

positions_obs : observable des positions des noeuds

arbre : l'arbre courant issu de l'algorithme

L2F2_Dequantification_App.mis_a_jour_branches — Method

mis_a_jour_branches(Observable{Int} x Int)::Nothing

Met à jour le nombre de branches affiché à l'écran au fur et à mesure de l'animation.

Paramètres:

observable du nombre de branches.

nouvelle valeur qu'on va mettre à jour.

L2F2_Dequantification_App.preparer_donnees — Method

preparer_donnes(Vector{Int16} x Ref{Union{Nothing, Vector{Int16}}} x Ref{Union{Nothing, Dict{Tuple{Int16,Int16}, Int}} x String)

La fonction calcule xQ, la série sous-quantifiée à 1 bit de x et P l'histogramme des couples. Elle les sauvegarde dans un dossier pour que l'utilisateur puisse les récupérer et les stocke dans leurs références respectives (xQref et Pref) pour lancer_animation

Paramètres:

la série x

référence de x

référence de P

dossier de sauvegarde

L2F2_Dequantification_App.relier_aretes! — Method

relier_aretes!(SimpleDiGraph x Noeud x Dict{Noeud,Int})::Nothing

Parcourt l'arbre en profondeur et relie chaque noeud à ses enfants dans le graphe.

Paramètres :

graphe : le graphe dans lequel on ajoute les arêtes

noeud : noeud courant à traiter

indices : dictionnaire associant chaque noeud à son identifiant dans le graphe

L2F2_Dequantification_App.telecharger_solutions — Method

telecharger_solutions(vector{String} x Observable{String})::Nothing

Télécharge les solutions dans un dossier choisi par l'utilisateur en format ZIP.

Paramètres:

liste des solutions générées

observable du message de succès en cas de réussite du téléchargement.

L2F2_Dequantification_App.telecharger_xQP — Method

telecharger_xQP(String x Observable{String})::Nothing

Copie xQ.dat et P.ppm depuis le dossier de l'application vers un dossier choisi par l'utilisateur.

Paramètres:

dossier de sauvearde de xQ et P dans l'application.

observable messagesuccesobs permettant d'afficher un message pour indiquer que le téléchargement a réussi.